+++ packages/iceworks-server/src/lib/adapter/dependency/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "dependency",
+  "events": [
+    {
+      "name": "project.dependency.list",
+      "handler": "project.dependency.list"
+    }
+  ]
+}
